+++ CHANGELOG.md
@@ -4,3 +4,9 @@
 * 6 categories
 * Optionally add keywords to recommend emojis
 * Material design and Cupertino mode
+
+
+## [0.0.2] - 21/04/2019
+
+* Removed NeverScrollableScrollPhysics class
+* Improved scroll performance